requirements.txt
transformers
torch
fastapi
uvicorn
pydantic
python-dotenv
PyYAML
gunicorn
loguru
